log: Show refs that point to each commit
This is another user convenience feature. Let's modify `log` to print refs next
to commits, so that for each commit that is printed we will additionally print
all refs that point to that commit.

First let's build a reverse lookup dictionary, from a commit OID to the refs
that reference it.

Then for each commit we will print the refs next to the OID.

%% METADATA
Id: annotate-log