show: Print diff of commit
Finally, instead of just printing the changed files in a commit, let's print the
actual diff (diff is the format with pluses and minuses at the beginning of the
line).

Hopefully you're already very familiar with the diff format because that's the
format I used to encode the changes you're reading here :)

We will create a function `diff_blobs()` that will take two blob OIDs and return
the diff between them. For each changed file, we will call `diff_blobs()` with the
OID of that file in the parent commit and the OID in the current commit.

Since the difference between the parent commit and the current commit is exactly
the change that was performed in the commit, we will effectively print all lines
changed by the commit.

For `diff_blobs()`, we're going to use an external Unix utility called "diff".
It receives two files as arguments and prints the differences between them in
diff format, just like we need. We'll need to read the blobs from the object
database and write them to a temporary file, which will be given to "diff"
utility.

Since "diff"'s output is a byte string, we will output it raw to stdout using
`sys.stdout.buffer.write()`.

(When invoking "diff" utility, I passed some extra options (--unified,
--show-c-function, --label) to make the diff look prettier. You can read about
these options in the man page of diff.)

And that's it! We can actually run `ugit show` on any commit and see the full
content of the commit. Try it!

%% METADATA
Id: show-diff