merge: Support multiple parents
Do you remember that the merge looked like this?

    o---o---o---o---o
         \           \
          --o---o-----o < HEAD
                ^
                some-branch

The HEAD commit is a commit that merges two commits together, and therefore the
HEAD commit has two parent commits. Right now ugit doesn't support that, so we
will extend the commit object to optionally support multiple parents.

The `Commit` object in *base.py* will now contain a list of parents rather than
a single parent.

The important modification is in `base.get_commit()`, which parses commit objects:
For each "parent" line that we encounter, we would append it to the list of
parents. So if the commit has one parent, the list will contain one element. If
the commit has multiple parents, the list will contain all parents. If the
commit has no parents (the first commit), the list will be empty.

The rest of the modifications are simple refactoring to support the new `Commit`
object.

%% METADATA
Id: support-multiple-parents