fetch: Download missing objects pointed by remote refs
In addition to fetching the refs, we're going to fetch the objects that the refs
points to. For that we will need some extra functions:

In data.py
==========

We will add `fetch_object_if_missing` to conditionally copy objects from a
remote repository by OID.

In base.py
==========

We will add `iter_objects_in_commits`, which will take a list of commit OIDs and
return all objects that are reachable from these commits. The code is a bit
large, but it mostly relies on `iter_commits_and_parents` to get a list of all
commits and then recursively iterates on the trees in each commit.

It's important to note that before accessing an OID we yield it, to give the
caller a chance to fetch it if it's missing.

In remote.py
============

We've done the heavy lifting in *data.py* and *base.py*, so in `fetch` we just
need to iterate over all objects and fetch the missing ones.

And that's it! We're done, `fetch` works! It's not the most efficient
implementation but it's very functional.

Now you can use `fetch` followed by a `merge` to synchronize your current
working branch to a remote repository.

%% METADATA
Id: fetch-remote-refs-objects