status: Show staged and non-staged modified files
To make user's life easier, let's extend the `status` command to display which
changed files will be committed and which files were changed but not going to be
committed.

The change is quite simple because it relies on existing infrastructure. To
display changed files that **are** going to be committed, we will compare index
to HEAD. To display changed files that **aren't** going to be committed we will
compare the index to the working tree.

Note that a file can be present in both groups - maybe we changed it, added it
to the index, changed it some more but didn't add the latest changes to the
index.

%% METADATA
Id: status-show-staged