write-tree: Write from the index instead of the working directory
Previously, `write_tree()` would scan the directory recursively, writing tree
objects for each directory it encounters. Now it will take its data from the
index. For this we will need to completely rewrite `write_tree()`.

First, we need the convert the index, which is a flat list of files, to a tree
of dictionaries. The index is stored as a flat list to make adding and removing
items easier and to make it compatible with our diff logic in *diff.py* (as we
will see later). However ugit stores the full tree in a hierarchy of tree
objects, so we need the data as a tree of dictionaries.

Then we will call `write_tree_recursive()` which is basically the same as the
old `write_tree()`, but instead of operating on the filesystem it will work with
the tree of dicts and write them to the objects store.

Cool! Now the index kind of works. From now on you must remember to `add` files
before committing them or else they won't be part of the commit.

%% METADATA
Id: write-tree-from-index